Importer changes
Added support for loading only the streams that contain the selected channels. This speeds up loading.

Instead of loading a truncated version with just the selected channels I create a sparse matrix which is populated only on the selected channels. The reason for this is that there is a problem with the ICA if the matrix is not full.